requirements.txt
hibachi-xyz>=0.1.14
python-dotenv>=1.0.0
requests>=2.31.0
certifi>=2024.8.30